% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movienetdata.R
\docType{package}
\name{movienetdata}
\alias{movienetdata}
\title{movienetdata: A data package for character interactions in popular films.}
\description{
This package contains character interaction data for 30 popular
  films, detailed below. The package is designed to play well with the
  \code{charinet} package for character interaction network analysis (under
  development), but the data can be used for any purpose.
}
\details{
The films for which data is available in this package include:
  \itemize{\item the first 18 films of the Marvel Cinematic Universe (from
  \emph{Iron Man} (2008) to \emph{Black Panther} (2018)); \item the 4 films
  in \emph{The Hunger Games} series (2012-2015); \item the 3 \emph{Tomb
  Raider} films, including the two \emph{Lara Croft: Tomb Raider} films
  (2001-2003) as well as the 2018 reboot; \item 5 miscellaneous other films
  (\emph{Frozen} (2013), \emph{The Breakfast Club} (1985), \emph{Star Wars:
  The Force Awakens} (2015), \emph{Alien} (1979), and \emph{Wonder Woman}
  (2017).}

  For full details of the films included in the package, run \preformatted{
  data(package = "movienetdata")} or consult the \code{?help} page for each
  dataset.

  Individual films can be loaded into the workspace like so:
  \preformatted{  data(marvel_01, package = "movienetdata")}

  Or by assigning them directly:
  \preformatted{  iron_man <- movienetdata::marvel_01}
}
\section{Citation}{
 If you find a use for the data, please either cite the
  package directly or the author's associated work, e.g.:

  Jones, Pete. 2020. "A social network analysis approach to examining
  gendered character positions in popular film narratives." Doctoral thesis,
  University of Manchester, UK.
}

\author{
Pete Jones
}
