% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thebreakfastclub}
\alias{thebreakfastclub}
\title{The Breakfast Club (1985).}
\format{
A list with three elements: \enumerate{ \item The event list (a
  matrix). Each row contains information on a line of dialogue, with columns
  corresponding to \code{eventID}, \code{sceneID}, \code{speakerID}, and
  dummy variables for each recipient. \item The node list (a data frame).
  Rows correspond to named speaking characters, with columns corresponding to
  variables for ID, name, total lines spoken (\code{nlines}), total times
  spoken to (\code{linesin}), and gender. \item The adjacency matrix. This is
  derived from the event list through aggregation and can be useful for
  static purposes such as visualising the network as a sociogram.}
}
\source{
Package author.
}
\usage{
thebreakfastclub
}
\description{
Character interaction data for the movie \emph{The Breakfast Club} (1985).
}
\references{
Jones, Pete. 2020. "A social network analysis approach to
  examining gendered character positions in popular film narratives."
  Doctoral thesis, University of Manchester, UK.
}
\keyword{datasets}
